backend/apigateway/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
redis==6.1.0
django-cors-headers==4.6.0
PyJWT==2.9.0
requests==2.32.3
celery==5.5.2
channels==4.2.2
websocket-client==1.8.0

